Virtual sites refactor
based on review and offline discusison
* pressure calculation moved into the VirtualSites implementation classes.
* virtual sites can decide how many pressure/stress contribs they want
* renamed stress/pressure contribution to "virtual_sites"  in core and py interface
* moved VirtualSites and derived classes into src/core/virtual_sites
* functions are in src/core/virtual_stes.?pp
* unneeded lb checking removed
* unneeded type checking in script interface removed
