+++ src/core/virtual_sites/VirtualSites.hpp
@@ -18,8 +18,8 @@
   You should have received a copy of the GNU General Public License
   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-#ifndef _VIRTUAL_SITES_HPP
-#define _VIRTUAL_SITES_HPP
+#ifndef VIRTUAL_SITES_VIRTUAL_SITES_HPP
+#define VIRTUAL_SITES_VIRTUAL_SITES_HPP
 
 
 /** \file virtual_sites.hpp
@@ -48,6 +48,10 @@ class VirtualSites {
     virtual void update(bool recalc_positions=true) const =0;
     /** Back-transfer forces (and torques) to non-virtual particles */
     virtual void back_transfer_forces_and_torques() const =0;
+    /** @brief Number of pressure contributions */
+    virtual int n_pressure_contribs() const {return 0;};
+    /** @brief Pressure contribution() */
+    virtual void pressure_and_stress_tensor_contribution(double* pressure, double* stress_tensor) const {};    
     /** @brief Enable/disable velocity calculations for vs */
     void set_have_velocity(const bool& v) { m_have_velocity=v; };
     const bool& have_velocity() const { return m_have_velocity; };
@@ -62,26 +66,5 @@ class VirtualSites {
    };
 
 
-   /** @brief Do-nothing virtual-sites implementation */
-   class VirtualSitesOff : public VirtualSites {
-    /** @brief Update positions and/or velocities of virtual sites 
-
-    * Velocities are only updated update_velocities() return true 
-    * @param recalc_positions can be used to skip the reculation of positions 
-    */
-    void update(bool recalc_positions=true) const override {};
-    /** Back-transfer forces (and torques) to non-virtual particles */
-    void back_transfer_forces_and_torques() const override {};
-    /** @brief Is a ghost communication needed after position updates */
-    bool need_ghost_comm_after_pos_update() const override { return false;} 
-    /** Is a ghost comm needed before a velocity update */
-    bool need_ghost_comm_before_vel_update() const override {return false;};
-    /** Is a ghost comm needed before back_transfer */
-    bool need_ghost_comm_before_back_transfer() const override {return false;};
-   };
-
-const std::shared_ptr<VirtualSites>& virtual_sites();
-
-void set_virtual_sites(std::shared_ptr<VirtualSites> v); 
 #endif
 #endif
