--- src/core/virtual_sites_com.hpp
@@ -1,46 +1,0 @@
-/*
-  Copyright (C) 2010,2011,2012,2013,2014,2015,2016 The ESPResSo project
-  Copyright (C) 2002,2003,2004,2005,2006,2007,2008,2009,2010
-    Max-Planck-Institute for Polymer Research, Theory Group
-
-  This file is part of ESPResSo.
-
-  ESPResSo is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  ESPResSo is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-#ifndef _VIRTUAL_SITES_COM_H
-#define _VIRTUAL_SITES_COM_H
-
-#include "config.hpp"
-
-#ifdef VIRTUAL_SITES_COM
-#include "particle_data.hpp"
-
-// The following three functions have to be provided by all implementations
-// of virtual sites
-// Update the vel/pos of the given virtual particle as defined by the real
-// particles in the same molecule
-void update_mol_pos_particle(Particle *);
-void update_mol_vel_particle(Particle *);
-
-// Distribute forces that have accumulated on virtual particles to the
-// associated real particles
-void distribute_mol_force();
-
-// Gets the (first) virtual particle of the same molecule as the given (real)
-// particle
-Particle *get_mol_com_particle(Particle *calling_p);
-
-#endif
-
-#endif
